new api for getting customer's distance

+++ app/src/main/java/com/hawkerapp/app/models/UserRequestData.kt
@@ -15,5 +15,5 @@ data class UserRequestData(
     @SerializedName("status")
     val status: String = "PENDING",
     @SerializedName("distance")
-    val distance: Long){
+    val distance: Double){
 }
